requirements.txt
alembic==1.13.1
annotated-types==0.6.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
editdistance==0.8.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fuzzywuzzy==0.18.0
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.0
Mako==1.3.2
MarkupSafe==2.1.5
mypy==1.9.0
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
SQLAlchemy==2.0.29
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
